Refactoring export lib
